requirements.txt
# Core dependencies
requests>=2.27.0
python-dotenv>=0.19.0
numpy>=1.24.0
pandas>=1.3.0
matplotlib>=3.5.0
scikit-learn>=1.2.0

# API integrations
alpaca-trade-api>=2.3.0

# Visualization
plotly>=5.5.0
seaborn>=0.12.0

# MCP LLM integration
openai>=1.0.0
langchain>=0.0.300
transformers>=4.30.0
torch>=2.0.0

yfinance>=0.2.12
ta==0.10.2
websockets==11.0.3
aiohttp==3.8.5
pydantic==2.3.0
tabulate>=0.8.9
joblib==1.3.2
tensorflow==2.13.0
xgboost==2.0.0

# Additional dependencies
websocket-client>=1.2.1
scipy>=1.7.0
beautifulsoup4>=4.10.0
tqdm>=4.62.0 